autotune: move occupancy lowering into the ROCm backend
Addresses review feedback on #785 (zhiding512): occupancy is a target-specific
concern that belongs in the backend, and the lowering chunk was an odd fit in
jit_function.py. It now lives in the ROCm backend, still consumed in exactly one
place (the working gpu.func-attribute mechanism); the dead gpu-module-to-binary
opts= path stays removed.

- base: BaseBackend.apply_occupancy_hints(module) -- no-op default.
- rocm: RocmBackend.apply_occupancy_hints() owns the gpu.func-attribute lowering;
  the helpers (_iter_gpu_kernel_funcs / _set_passthrough / _set_occupancy_attrs /
  _resolve_occupancy_hint / _unmatched_occupancy_hint_keys /
  _apply_occupancy_compile_hints) move here verbatim. `ir` is imported lazily so
  the backend module stays importable for backend discovery without the compiled
  _mlir bindings.
- jit_function: MlirCompiler.compile calls backend.apply_occupancy_hints(module);
  _stable_hint_repr (a compile-hint cache-key helper, not occupancy) stays.
- tests: import the occupancy helpers from flydsl.compiler.backends.rocm.

No behavior change (same attributes set at the same point in compile). Refs #785.
