+++ python/flydsl/compiler/jit_function.py
@@ -787,155 +787,6 @@ def _run_pipeline(module: ir.Module, fragments: list, *, verifier: bool, print_a
             raise DSLCompileError(str(exc), diagnostics=diags) from exc
 
 
-def _iter_gpu_kernel_funcs(module: ir.Module):
-    """Yield the entry-point kernel ``gpu.func`` ops (those carrying the
-    ``gpu.kernel`` attribute) inside every ``gpu.module`` of *module*.
-
-    Non-kernel device helpers are skipped -- occupancy hints only apply to entry
-    points, and a module may hold both."""
-    for top in module.body.operations:
-        if top.operation.name != "gpu.module":
-            continue
-        for op in top.regions[0].blocks[0].operations:
-            if op.operation.name == "gpu.func" and "gpu.kernel" in op.attributes:
-                yield op
-
-
-def _set_passthrough(func_op, key: str, value: str) -> None:
-    """Set an LLVM ``passthrough`` ``[key, value]`` entry on a kernel func,
-    replacing any existing entry with the same key.
-
-    ``convert-gpu-to-rocdl`` copies the ``passthrough`` discardable attr from the
-    gpu.func onto the lowered llvm.func, where the LLVM emitter turns each entry
-    into a function attribute -- bridging AMDGPU function attributes the ROCDL
-    dialect does not translate natively (e.g. ``amdgpu-num-vgpr``). This is
-    related to, but distinct from, ``FlyROCDLClusterAttrPass``: that pass copies
-    a ``rocdl.cluster_dims`` attr and *synthesizes* the passthrough on the
-    llvm.func in a post-lowering pass, whereas here the ``passthrough`` is set
-    directly on the gpu.func and carried through as-is. Preserves unrelated
-    entries but must not leave a duplicate key (duplicate LLVM function
-    attributes are ill-defined).
-    """
-
-    def _entry_key(e):
-        # Key/value entries are 2-element arrays [key, value]; unit attributes
-        # (e.g. "nounwind") are bare strings and have no key to match.
-        try:
-            arr = ir.ArrayAttr(e)
-            return ir.StringAttr(arr[0]).value if len(arr) else None
-        except (ValueError, TypeError):
-            return None
-
-    entry = ir.ArrayAttr.get([ir.StringAttr.get(key), ir.StringAttr.get(value)])
-    existing = func_op.attributes["passthrough"] if "passthrough" in func_op.attributes else None
-    kept = [e for e in existing if _entry_key(e) != key] if existing is not None else []
-    func_op.attributes["passthrough"] = ir.ArrayAttr.get(kept + [entry])
-
-
-def _set_occupancy_attrs(func_op, *, waves_per_eu=None, maxnreg=None) -> None:
-    """Single source of truth for writing occupancy knobs onto a kernel
-    ``gpu.func`` as the attributes the ROCDL / LLVM lowering actually honors.
-
-    There is exactly one working mechanism for occupancy -- these ``gpu.func``
-    attributes:
-
-      - ``waves_per_eu`` -> ``rocdl.waves_per_eu`` (translated by convert-gpu-to-rocdl)
-      - ``maxnreg``      -> ``amdgpu-num-vgpr`` LLVM passthrough (no native ROCDL attr)
-
-    The autotune compile-hint path (:func:`_apply_occupancy_compile_hints`)
-    routes here, and hand-authored kernels that set ``rocdl.waves_per_eu`` via
-    ``value_attrs`` land on the same attribute. Passing these knobs through
-    ``gpu-module-to-binary opts=`` does NOT work -- the AMDGPU backend drops them
-    silently. A value of 0 / None means "leave it to the compiler". Must be
-    called with *func_op*'s MLIR context active.
-    """
-    if waves_per_eu:
-        i32 = ir.IntegerType.get_signless(32)
-        func_op.attributes["rocdl.waves_per_eu"] = ir.IntegerAttr.get(i32, int(waves_per_eu))
-    if maxnreg:
-        _set_passthrough(func_op, "amdgpu-num-vgpr", str(int(maxnreg)))
-
-
-def _resolve_occupancy_hint(hint, sym_name: str):
-    """Resolve one occupancy compile-hint for a single kernel entry point.
-
-    A hint is either a scalar (applied uniformly to every ``gpu.kernel``) or a
-    ``{sym_name: value}`` mapping (per-kernel; kernels absent from the map are
-    left to the compiler -- i.e. resolve to ``None``).
-    """
-    if isinstance(hint, Mapping):
-        return hint.get(sym_name)
-    return hint
-
-
-def _unmatched_occupancy_hint_keys(hints: dict, present: set) -> dict:
-    """Map ``knob -> [keys]`` for per-kernel occupancy hints whose keys name no
-    kernel in *present* (the ``sym_name``s of the module's entry kernels).
-
-    A ``{sym_name: value}`` hint with a stale/typo'd key would otherwise be a
-    silent no-op -- the exact "silently ignored" failure this occupancy rework
-    set out to eliminate -- so the caller warns on any leftovers.
-    """
-    out = {}
-    for knob, hint in hints.items():
-        if isinstance(hint, Mapping):
-            missing = sorted(k for k in hint if k not in present)
-            if missing:
-                out[knob] = missing
-    return out
-
-
-def _apply_occupancy_compile_hints(module: ir.Module) -> None:
-    """Lower the autotuner's occupancy compile-hints onto each kernel gpu.func.
-
-    ``Config.compiler_opts()`` surfaces occupancy knobs (``waves_per_eu``,
-    ``maxnreg``) as thread-local ``compile_hints``. They only take effect as
-    kernel function attributes, so this walks the entry-point kernels and writes
-    them via :func:`_set_occupancy_attrs` -- the single occupancy mechanism. The
-    dead ``gpu-module-to-binary opts=`` route (silently ignored by the AMDGPU
-    backend) is intentionally not used.
-
-    Per-kernel vs uniform: each hint may be a scalar ``int`` -- applied to
-    *every* ``gpu.kernel`` entry point (the common case, e.g. single-kernel
-    launchers like rmsnorm) -- or a ``{sym_name: int}`` mapping resolved per
-    kernel func against its ``sym_name`` (kernels absent from the map are left
-    to the compiler; a map key naming no kernel is warned about, not silently
-    dropped). The mapping form lets a multi-kernel ``@jit`` launcher scope
-    occupancy per entry kernel via ``CompilationContext.compile_hints``. Note:
-    ``Config`` can carry such a mapping, but the autotune *search* still
-    enumerates scalar knobs only, so autotune-driven *independent* per-kernel
-    tuning is not wired end-to-end yet -- the search would need to explore the
-    mappings (deferred; see PR #785).
-    """
-    from .kernel_function import CompilationContext
-
-    hints = CompilationContext.get_compile_hints()
-    waves_per_eu = hints.get("waves_per_eu")
-    maxnreg = hints.get("maxnreg")
-    if not waves_per_eu and not maxnreg:
-        return
-    seen = set()
-    with module.context:
-        for func_op in _iter_gpu_kernel_funcs(module):
-            sym_name = ir.StringAttr(func_op.attributes["sym_name"]).value
-            seen.add(sym_name)
-            _set_occupancy_attrs(
-                func_op,
-                waves_per_eu=_resolve_occupancy_hint(waves_per_eu, sym_name),
-                maxnreg=_resolve_occupancy_hint(maxnreg, sym_name),
-            )
-    for knob, missing in _unmatched_occupancy_hint_keys(
-        {"waves_per_eu": waves_per_eu, "maxnreg": maxnreg}, seen
-    ).items():
-        log().warning(
-            "occupancy hint %r targets kernel(s) %s not present in the module (kernels: %s); "
-            "those entries were ignored -- check for a typo or stale sym_name.",
-            knob,
-            missing,
-            sorted(seen),
-        )
-
-
 def _stable_hint_repr(value) -> str:
     """Order-independent string form of a compile-hint value for the cache key.
 
@@ -966,7 +817,7 @@ class MlirCompiler:
         backend = get_backend(arch=arch)
 
         module = ir.Module.parse(module.operation.get_asm(enable_debug_info=env.debug.enable_debug_info))
-        _apply_occupancy_compile_hints(module)
+        backend.apply_occupancy_hints(module)
         cfg = _pipeline_fragments_for_mode(backend)
         fragments = cfg.fragments
         pre_binary_fragments = cfg.pre_binary
